requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
langchain_chroma
pandas
langchain-groq
pymupdf
mysql-connector-python
SQLAlchemy
validators==0.28.1
unstructured
pytube
numexpr
huggingface_hub
langchain_core
streamlit
